13 support tags for series

+++ app/com/playframework/cric/models/ManOfTheSeries.java
@@ -15,7 +15,7 @@ public class ManOfTheSeries extends Model {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private Long seriesId;
+    private Integer seriesId;
     private Long playerId;
 }
