+++ app/com/playframework/cric/repositories/SeriesTeamsMapRepository.java
@@ -11,21 +11,21 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class SeriesTeamsMapRepository {
-    public void create(Long seriesId, List<Long> teamIds) {
+    public void create(Integer seriesId, List<Long> teamIds) {
         List<SeriesTeamsMap> seriesTeamsMaps = teamIds.stream().map(teamId -> new SeriesTeamsMap(null, seriesId, teamId)).collect(Collectors.toList());
         DB.saveAll(seriesTeamsMaps);
     }
 
-    public List<SeriesTeamsMap> getBySeriesIds(List<Long> seriesIds) {
+    public List<SeriesTeamsMap> getBySeriesIds(List<Integer> seriesIds) {
         return DB.find(SeriesTeamsMap.class).where().in("seriesId", seriesIds).findList();
     }
 
-    public void delete(Long seriesId, List<Long> teamIds) {
+    public void delete(Integer seriesId, List<Long> teamIds) {
         List<SeriesTeamsMap> seriesTeamsMaps = DB.find(SeriesTeamsMap.class).where().eq("seriesId", seriesId).in("teamId", teamIds).findList();
         DB.deleteAll(seriesTeamsMaps);
     }
 
-    public void remove(Long seriesId)
+    public void remove(Integer seriesId)
     {
         DB.deleteAll(getBySeriesIds(Collections.singletonList(seriesId)));
     }
