+++ app/com/playframework/cric/services/SeriesTeamsMapService.java
@@ -14,19 +14,19 @@ public class SeriesTeamsMapService {
         this.seriesTeamsMapRepository = seriesTeamsMapRepository;
     }
 
-    public void create(Long seriesId, List<Long> teamIds) {
+    public void create(Integer seriesId, List<Long> teamIds) {
         seriesTeamsMapRepository.create(seriesId, teamIds);
     }
 
-    public List<SeriesTeamsMap> getBySeriesIds(List<Long> seriesIds) {
+    public List<SeriesTeamsMap> getBySeriesIds(List<Integer> seriesIds) {
         return seriesTeamsMapRepository.getBySeriesIds(seriesIds);
     }
 
-    public void delete(Long seriesId, List<Long> teamIds) {
+    public void delete(Integer seriesId, List<Long> teamIds) {
         seriesTeamsMapRepository.delete(seriesId, teamIds);
     }
 
-    public void remove(Long seriesId)
+    public void remove(Integer seriesId)
     {
         seriesTeamsMapRepository.remove(seriesId);
     }
